pyproject.toml
[tool.poetry]
name = "ocr-server"
version = "0.1.0"
description = ""
authors = ["bonni-e <trial1bloom@gmail.com>"]
readme = "README.md"
packages = [{include = "ocr_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pyuploadcare = "^4.1.0"
mysqlclient = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
